refactoring details handlers
