+++ jupyterlab_bigquery/jupyterlab_bigquery/__init__.py
@@ -1,8 +1,7 @@
 """Initialize server endpoints for extension"""
 from notebook.utils import url_path_join
 
-from jupyterlab_bigquery.list_items_handler import Handlers
-from jupyterlab_bigquery.details_handler import DatasetDetailsHandler, TablePreviewHandler, TableDetailsHandler, ViewDetailsHandler, ModelDetailsHandler, TrainingRunDetailsHandler
+from jupyterlab_bigquery.create_handler import Handlers
 from jupyterlab_bigquery.version import VERSION
 from jupyterlab_bigquery.query_history_handler import QueryHistoryHandler, GetQueryDetailsHandler
 from jupyterlab_bigquery.pagedAPI_handler import PagedQueryHandler
@@ -38,12 +37,6 @@ def load_jupyter_server_extension(nb_server_app):
       [
           # TODO(cbwilkes): Add auth checking if needed.
           # (url_path_join(gcp_v1_endpoint, auth'), AuthHandler)
-          make_endpoint('datasetdetails', DatasetDetailsHandler),
-          make_endpoint('tabledetails', TableDetailsHandler),
-          make_endpoint('tablepreview', TablePreviewHandler),
-          make_endpoint('viewdetails', ViewDetailsHandler),
-          make_endpoint('modeldetails', ModelDetailsHandler),
-          make_endpoint('trainingRunDetails', TrainingRunDetailsHandler),
           make_endpoint('query', PagedQueryHandler),
           make_endpoint('projectQueryHistory', QueryHistoryHandler),
           make_endpoint('getQueryDetails', GetQueryDetailsHandler)
