+++ jupyterlab_bigquery/jupyterlab_bigquery/create_handler/__init__.py
@@ -1,0 +1,3 @@
+import jupyterlab_bigquery.list_items_handler.list_items_handler
+import jupyterlab_bigquery.details_handler.details_handler
+from .create_handler import Handlers
