--- jupyterlab_bigquery/jupyterlab_bigquery/details_handler/__init__.py
@@ -1,3 +1,0 @@
-from .details_handler import DatasetDetailsHandler, TableDetailsHandler, TablePreviewHandler,\
-   ViewDetailsHandler, ModelDetailsHandler, format_preview_fields, format_preview_rows,\
-      parallel_format_preview_rows
