--- jupyterlab_bigquery/jupyterlab_bigquery/list_items_handler/__init__.py
@@ -1,3 +1,0 @@
-"""Create the handlers for listing items and save in handlers{}"""
-import jupyterlab_bigquery.list_items_handler.list_items_handler
-from .create_handler import Handlers
