+++ jupyterlab_bigquery/jupyterlab_bigquery/list_items_handler/list_items_handler.py
@@ -1,9 +1,8 @@
 # Lint as: python3
 """Decorators that create handlers for the data tree"""
 
-import jupyterlab_bigquery.list_items_handler.create_handler
 from jupyterlab_bigquery.list_items_handler.service import BigQueryService
-from jupyterlab_bigquery.list_items_handler.create_handler import _handler
+from jupyterlab_bigquery.create_handler.create_handler import _handler
 
 SCOPE = ("https://www.googleapis.com/auth/cloud-platform",)
